python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nuclei-sdk"
version = "1.0.0"
description = "Asynchronous Python client for the Nuclei vulnerability scanner"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    { name = "D. Sanjai Kumar", email = "bughunterz0047@gmail.com" }, { name = "Gokul.V", email = "th3gokul@gmail.com"} , {name="Pugalarasan K", email="0xpugal@gmail.com"}
]
maintainers = [
    { name = "D. Sanjai Kumar", email = "bughunterz0047@gmail.com" }, { name = "Gokul.V", email = "th3gokul@gmail.com"} , {name="Pugalarasan K", email="0xpugal@gmail.com"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Security",
]
keywords = ["nuclei", "security", "scanner", "vulnerability", "python-nuclei", "vulnerability-scanner", "python-nuclei-sdk"]

[project.optional-dependencies]
templates = ["pyyaml>=6.0"]

[project.urls]
Homepage = "https://github.com/RevoltSecurities/nuclei-sdk"
Repository = "https://github.com/RevoltSecurities/nuclei-sdk"
Issues = "https://github.com/RevoltSecurities/nuclei-sdk/issues"

[tool.setuptools.packages.find]
include = ["nucleisdk*"]
